Add files via upload
Generate a signal of the form
s(n) = Asin (2πf1n/Fs) + Bsin(2πf2n/Fs)
where Fs is sampling frequency and the frequencies f1 and f2 can be chosen as desired.
Choose Fs so that the sinusoid appears smooth. A noise w(n) is added to the signal to produce the noisy signal
d(n) = s(n) + w(n)
The noise is a white uniformly distributed sequence with zero mean. The SNR of this signal should about −7 dB. The desired signal (or primary signal) is d(n).
The reference signal is d(n − 1) and is applied to the input of an N-tap FIR filter that produces the output y(n), which should be the estimate of the signal.
Use the LMS algorithm for adaptation. Choose N and step size appropriately for good results (hint: Step size = 0.0005. In this case, the filter order is 64. some parameters are A = 0.5, B = 1, f1 = 100Hz and f2 = 150Hz. Sampling frequency Fs = 500Hz.)
1. Plot the MSE for an ensemble of 100 runs.
2. Plot a few cycles of s(n) for one run.
3. Plot a few cycles of y(n) after convergence for one run.
4. Plot the impulse response and the frequency response of the filter after convergence (one run).
5. Find the SNR of the filtered signal.